# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/notifications": {
      "get": {
        "operationId": "NotificationController_list",
        "parameters": [
          {
            "name": "isRead",
            "required": false,
            "in": "query",
            "description": "Filter by read status (true/false)",
            "schema": {
              "example": false,
              "type": "boolean"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated notification list",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "allOf": [
                            {
                              "$ref": "#/components/schemas/PaginatedDataDto"
                            },
                            {
                              "properties": {
                                "items": {
                                  "type": "array",
                                  "items": {
                                    "$ref": "#/components/schemas/NotificationResponseDto"
                                  }
                                }
                              }
                            }
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List the current user's notifications (newest first). Filter by isRead and paginate with page/limit.",
        "tags": [
          "Notification"
        ]
      }
    },
    "/api/notifications/read-all": {
      "put": {
        "operationId": "NotificationController_markAllAsRead",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "null",
                          "nullable": true
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mark all of the current user's notifications as read",
        "tags": [
          "Notification"
        ]
      }
    },
    "/api/notifications/{id}/read": {
      "put": {
        "operationId": "NotificationController_markAsRead",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "null",
                          "nullable": true
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mark a single notification as read",
        "tags": [
          "Notification"
        ]
      }
    },
    "/api/notifications/settings": {
      "get": {
        "operationId": "NotificationController_getSettings",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/NotificationSettingsResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get the current user's notification preferences",
        "tags": [
          "Notification"
        ]
      },
      "post": {
        "operationId": "NotificationController_saveSettings",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveNotificationSettingsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/NotificationSettingsResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Save (upsert) the current user's notification preferences",
        "tags": [
          "Notification"
        ]
      }
    },
    "/api/notifications/test": {
      "post": {
        "operationId": "NotificationController_sendTest",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "null",
                          "nullable": true
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Send a test push to the current user's devices (dev/QA only — disabled in production unless PUSH_TEST_ENABLED=true)",
        "tags": [
          "Notification"
        ]
      }
    },
...
    "/api/dashboard/summary": {
      "get": {
        "operationId": "DashboardController_getSummary",
        "parameters": [
          {
            "name": "month",
            "required": false,
            "in": "query",
            "description": "Month (1–12). Must be supplied together with `year`.",
            "schema": {
              "minimum": 1,
              "maximum": 12,
              "example": 5,
              "type": "number"
            }
          },
          {
            "name": "year",
            "required": false,
            "in": "query",
            "description": "Year (e.g. 2026). Must be supplied together with `month`.",
            "schema": {
              "minimum": 1970,
              "maximum": 9999,
              "example": 2026,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/SummaryResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Financial overview: current balance + income/expense totals. Supply month+year together to scope to a single month, or omit both for all-time.",
        "tags": [
          "Dashboard"
        ]
      }
    },
    "/api/dashboard/recent": {
      "get": {
        "operationId": "DashboardController_getRecent",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of recent transactions to return",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/RecentTransactionDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Most recent transactions for the current user (newest first). Default limit 10, max 100.",
        "tags": [
          "Dashboard"
        ]
      }
    },
    "/api/devices": {
      "post": {
        "operationId": "DevicesController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDeviceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "null",
                          "nullable": true
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Register or refresh this device's FCM token for push notifications. Call on login and whenever Firebase issues a new token.",
        "tags": [
          "Device"
        ]
      }
    },
    "/api/devices/unregister": {
      "post": {
        "operationId": "DevicesController_unregister",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnregisterDeviceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "null",
                          "nullable": true
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove this device's FCM token (call on logout) so it stops receiving the user's push notifications.",
        "tags": [
          "Device"
        ]
      }
    }
  },
  "info": {
    "title": "Sonsam Saving API",
    "description": "Personal Finance Tracker — REST API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ApiResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 200,
            "description": "HTTP status code"
          },
          "message": {
            "type": "string",
            "example": "Success",
            "description": "Human-readable message"
          },
          "data": {
            "type": "object",
            "description": "Response payload (null when there is no body)"
          }
        },
        "required": [
          "statusCode",
          "message",
          "data"
        ]
      },
...
      "PaginationDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "example": 1
          },
          "limit": {
            "type": "number",
            "example": 20
          },
          "total": {
            "type": "number",
            "example": 45
          }
        },
        "required": [
          "page",
          "limit",
          "total"
        ]
      },
      "PaginatedDataDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "array"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationDto"
          }
        },
        "required": [
          "items",
          "pagination"
        ]
      },
...
      "NotificationResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1",
            "description": "Notification ID"
          },
          "title": {
            "type": "string",
            "example": "Transfer Received"
          },
          "message": {
            "type": "string",
            "example": "You received 20.00 from John Doe"
          },
          "isRead": {
            "type": "boolean",
            "example": false
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-05-10T14:00:00.000Z"
          }
        },
        "required": [
          "id",
          "title",
          "message",
          "isRead",
          "createdAt"
        ]
      },
      "NotificationSettingsResponseDto": {
        "type": "object",
        "properties": {
          "transferAlert": {
            "type": "boolean",
            "example": true
          },
          "monthlyReport": {
            "type": "boolean",
            "example": true
          },
          "lowBalanceAlert": {
            "type": "boolean",
            "example": true
          },
          "lowBalanceThreshold": {
            "type": "string",
            "example": "50.00",
            "description": "Decimal stored as string for precise client-side handling"
          }
        },
        "required": [
          "transferAlert",
          "monthlyReport",
          "lowBalanceAlert",
          "lowBalanceThreshold"
        ]
      },
      "SaveNotificationSettingsDto": {
        "type": "object",
        "properties": {
          "transferAlert": {
            "type": "boolean",
            "example": true,
            "description": "Receive a notification when a transfer is received"
          },
          "monthlyReport": {
            "type": "boolean",
            "example": true,
            "description": "Receive a monthly summary report"
          },
          "lowBalanceAlert": {
            "type": "boolean",
            "example": true,
            "description": "Receive an alert when balance falls below threshold"
          },
          "lowBalanceThreshold": {
            "type": "number",
            "example": 50,
            "description": "Threshold (in account currency) below which the low-balance alert fires"
          }
        },
        "required": [
          "transferAlert",
          "monthlyReport",
          "lowBalanceAlert",
          "lowBalanceThreshold"
        ]
      },
...
      "SummaryResponseDto": {
        "type": "object",
        "properties": {
          "totalBalance": {
            "type": "string",
            "example": "730.00",
            "description": "User's current account balance (lifetime, not period-scoped)"
          },
          "totalIncome": {
            "type": "string",
            "example": "1000.00",
            "description": "Sum of all income transactions in the period (or all-time if no period given). Includes transfer-in."
          },
          "totalExpense": {
            "type": "string",
            "example": "270.00",
            "description": "Sum of all expense transactions in the period (or all-time if no period given). Includes transfer-out."
          },
          "period": {
            "type": "string",
            "example": "May 2026",
            "description": "Human-readable period label, or \"All time\" when no month/year is supplied"
          }
        },
        "required": [
          "totalBalance",
          "totalIncome",
          "totalExpense",
          "period"
        ]
      },
      "RecentTransactionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "10",
            "description": "Transaction ID"
          },
          "amount": {
            "type": "string",
            "example": "15.50",
            "description": "Amount as decimal string"
          },
          "type": {
            "type": "string",
            "enum": [
              "income",
              "expense"
            ],
            "example": "expense"
          },
          "source": {
            "type": "string",
            "enum": [
              "manual",
              "transfer"
            ],
            "example": "manual"
          },
          "category": {
            "type": "string",
            "example": "Food",
            "description": "Category name"
          },
          "date": {
            "type": "string",
            "example": "2026-05-03",
            "description": "Transaction date (YYYY-MM-DD)"
          },
          "note": {
            "type": "string",
            "example": "Lunch",
            "nullable": true
          }
        },
        "required": [
          "id",
          "amount",
          "type",
          "source",
          "category",
          "date",
          "note"
        ]
      },
      "RegisterDeviceDto": {
        "type": "object",
        "properties": {
          "fcmToken": {
            "type": "string",
            "example": "fGc1...long-fcm-token",
            "description": "The FCM registration token issued to this device by Firebase"
          },
          "platform": {
            "type": "string",
            "enum": [
              "android",
              "ios"
            ],
            "example": "android"
          }
        },
        "required": [
          "fcmToken",
          "platform"
        ]
      },
      "UnregisterDeviceDto": {
        "type": "object",
        "properties": {
          "fcmToken": {
            "type": "string",
            "example": "fGc1...long-fcm-token",
            "description": "The FCM token to remove so this device stops receiving pushes"
          }
        },
        "required": [
          "fcmToken"
        ]
      }
    }
  }
}